PRL-5248: Update method to check for response document
