+++ src/main/steps/common/task-list/components/notification-banner/utils.test.ts
@@ -1,14 +1,8 @@
 import { CaseWithId } from '../../../../../app/case/case';
-import { Applicant, CaseType, Respondent, State } from '../../../../../app/case/definition';
+import { Applicant, CaseType, State } from '../../../../../app/case/definition';
 import { UserDetails } from '../../../../../app/controller/AppRequest';
 
-import {
-  BannerNotification,
-  hasResponseBeenSubmitted,
-  isApplicantLIPServingRespondent,
-  isPrimaryApplicant,
-  notificationBanner,
-} from './utils';
+import { BannerNotification, isApplicantLIPServingRespondent, isPrimaryApplicant, notificationBanner } from './utils';
 
 describe('notification Banner', () => {
   const data = {
@@ -96,80 +90,6 @@ describe('notification Banner', () => {
         ...data,
         state: 'PREPARE_FOR_HEARING_CONDUCT_HEARING',
       } as Partial<CaseWithId>)
-    ).toBe(false);
-  });
-
-  test('hasResponseBeenSubmitted should return true if response document is present', () => {
-    expect(
-      hasResponseBeenSubmitted(
-        {
-          citizenDocuments: [
-            {
-              partyId: '1',
-              partyName: null,
-              partyType: 'respondent',
-              categoryId: 'respondentApplication',
-              uploadedBy: 'test user',
-              uploadedDate: '2024-03-11T16:24:33.122506',
-              reviewedDate: '2024-03-11T16:24:33.122506',
-              document: {
-                document_url: 'MOCK_DOCUMENT_URL',
-                document_binary_url: 'MOCK_DOCUMENT_BINARY_URL',
-                document_filename: 'MOCK_FILENAME',
-                document_hash: null,
-                category_id: 'respondentApplication',
-                document_creation_date: '2024-03-11T16:24:33.122506',
-              },
-              documentWelsh: null,
-            },
-          ],
-        } as unknown as CaseWithId,
-        {
-          id: '1',
-          value: {
-            user: {
-              idamId: '1',
-            },
-          },
-        } as Respondent
-      )
-    ).toBe(true);
-  });
-
-  test('hasResponseBeenSubmitted should return false if response document is notpresent', () => {
-    expect(
-      hasResponseBeenSubmitted(
-        {
-          citizenDocuments: [
-            {
-              partyId: '1',
-              partyName: null,
-              partyType: 'respondent',
-              categoryId: 'positionStatements',
-              uploadedBy: 'test user',
-              uploadedDate: '2024-03-11T16:24:33.122506',
-              reviewedDate: '2024-03-11T16:24:33.122506',
-              document: {
-                document_url: 'MOCK_DOCUMENT_URL',
-                document_binary_url: 'MOCK_DOCUMENT_BINARY_URL',
-                document_filename: 'MOCK_FILENAME',
-                document_hash: null,
-                category_id: 'positionStatements',
-                document_creation_date: '2024-03-11T16:24:33.122506',
-              },
-              documentWelsh: null,
-            },
-          ],
-        } as unknown as CaseWithId,
-        {
-          id: '1',
-          value: {
-            user: {
-              idamId: '1',
-            },
-          },
-        } as Respondent
-      )
     ).toBe(false);
   });
 });
